Added some casts to reduce warnings from -Wconversion.
git-svn-id: https://svn.r-project.org/R/trunk@55416 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ src/appl/cpoly.c
@@ -156,7 +156,7 @@ void R_cpolyroot(double *opr, double *opi, int *degree,
     if (nn == 1) return;
 
     /* Use a single allocation as these as small */
-    tmp = (double *) R_alloc(10*nn, sizeof(double));
+    tmp = (double *) R_alloc((size_t) (10*nn), sizeof(double));
     pr = tmp; pi = tmp + nn; hr = tmp + 2*nn; hi = tmp + 3*nn;
     qpr = tmp + 4*nn; qpi = tmp + 5*nn; qhr = tmp + 6*nn; qhi = tmp + 7*nn;
     shr = tmp + 8*nn; shi = tmp + 9*nn;
